Add hello world console application
Above is the commit "subject" line. It should be kept short,
but still descriptive. If you need to add more details include
them on a newline below. It's also common practice to keep the
lines about as wide as the text below.  This ensures that the summary
can be viewed properly when running git log.